Small refactorings to allow for RenderTarget support for WebGPU
